+++ ai_sapiens_sim2real/src/teleop_devtools/dualsense/dualsense_teleop_ui.cpp
@@ -14,7 +14,7 @@
 //
 // Author: Kiwoong Park
 
-#include "ai_sapiens_sim2real/dualsense_teleop/dualsense_teleop_ui.hpp"
+#include "ai_sapiens_sim2real/teleop_devtools/dualsense/dualsense_teleop_ui.hpp"
 
 #include <algorithm>
 #include <cmath>
