+++ ai_sapiens_sim2real/src/teleop_devtools/dualsense/dualsense_teleop_ui_node.cpp
@@ -27,7 +27,7 @@
 #include <rclcpp/rclcpp.hpp>
 #include <yaml-cpp/yaml.h>  // NOLINT(build/include_order)
 
-#include "ai_sapiens_sim2real/dualsense_teleop/dualsense_teleop_ui.hpp"
+#include "ai_sapiens_sim2real/teleop_devtools/dualsense/dualsense_teleop_ui.hpp"
 #include "ai_sapiens_sim2real/plugins/teleop_input/dualsense_teleop_input_plugin.hpp"
 
 namespace ai_sapiens_sim2real
